Decompose game-simulator monolith + introduce views layer (Steps 1-3)
Split the ~184KB game-simulator.js monolith into focused domain modules under
src/core/simulation/ following the ZenGM worker/core pattern, and add a
worker/views-style data-preparation layer under src/views/.

- src/core/simulation/: playExecution, driveEngine, scoreKeeper,
  statAccumulator, clockManager, injuryResolver, gameSummaryBuilder, and an
  index.js orchestrator that is the only module wiring the domains together.
  game-simulator.js is now a thin re-export shim (4KB, was 184KB).
- AUDIT.md maps every responsibility of the original file to its new home.
- Simulation math is unchanged: the seeded Utils PRNG call order is preserved
  byte-for-byte, so play resolution is statistically identical. All existing
  Vitest tests (2471) still pass.
- src/views/: draftView, rosterView, gameResultView, standingsView — pure
  functions returning plain serializable view-models (no React/JSX/hooks).
- Draft.jsx now consumes prepareDraftView(league) instead of reading raw
  league state directly.
